Allow DH_set0_key() to set priv_key without pub_key
This is useful when reconstructing a DH from a saved private key. Once
we can create a DH with the private key (and the DH parameters), the
corresponding public key can be computed with DH_generate_key(). In
older versions, we could simply set DH.priv_key directly, however since
libcrypto was made opaque, we can no longer set directly.